Create a root App compoenent
