+++ src/static/index.html
@@ -9,7 +9,7 @@
   </head>
   <body>
     <input id="fileInput" type="file" accept="text/json" />
-    <div id="menuBarRoot"></div>
+    <div id="root"></div>
     <button id="menubarOpenButton" class="iconButton"></button>
     <div id="toolbar">
       <button id="deleteTaskButton" class="iconButton"></button>
